Moved all disease behaviour into the behaviour folder
