+++ Protecs/src/main/java/uk/ac/ucl/protecs/behaviours/InfectiousBehaviourFramework.java
@@ -1,8 +1,9 @@
-package uk.ac.ucl.protecs.objects.diseases;
+package uk.ac.ucl.protecs.behaviours;
 
 
 
 import uk.ac.ucl.protecs.objects.*;
+import uk.ac.ucl.protecs.objects.diseases.Infection;
 import uk.ac.ucl.protecs.sim.*;
 import sim.engine.Steppable;
 import swise.behaviours.BehaviourFramework;
