+++ Protecs/src/test/java/uk/ac/ucl/protecs/sim/CovidTestingTesting.java
@@ -7,7 +7,7 @@ import uk.ac.ucl.protecs.helperFunctions.helperFunctions.NodeOption;
 import org.junit.Test;
 
 import uk.ac.ucl.protecs.objects.Person;
-import uk.ac.ucl.protecs.objects.diseases.CoronavirusBehaviourFramework.CoronavirusBehaviourNodeTitle;
+import uk.ac.ucl.protecs.behaviours.CoronavirusBehaviourFramework.CoronavirusBehaviourNodeTitle;
 import uk.ac.ucl.protecs.objects.diseases.Infection;
 import uk.ac.ucl.protecs.sim.WorldBankCovid19Sim.DISEASE;
